# src/api/docs/swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "description": "WSES Api",
    "version": "1.0.0",
    "title": "WSES API"
  },
  "servers": [
    {
      "url": "http://192.168.0.120:5003/",
      "description": "Development server"
    },
    {
      "url": "https://mymeetingdesk.com:7410/",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/get-overview": {
      "get": {
        "summary": "Get overview",
        "tags": ["Overview"],
        "operationId": "getOverview",

        "description": "Get overview",
        "parameters": [
          {
            "$ref": "#/components/parameters/headerLanguage"
          },
          {
            "$ref": "#/components/parameters/headerPlatform"
          },
          {
            "$ref": "#/components/parameters/headerVersion"
          }
        ],
        "responses": {
          "200": {
            "description": "Get overview",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ApiResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          }
        }
      }
    },
    "/api/get-overview-list": {
      "get": {
        "summary": "Get overview List",
        "tags": ["Overview"],
        "operationId": "getOverviewList",

        "description": "Get overview list",
        "parameters": [
          {
            "$ref": "#/components/parameters/headerLanguage"
          },
          {
            "$ref": "#/components/parameters/headerPlatform"
          },
          {
            "$ref": "#/components/parameters/headerVersion"
          },
          {
            "in": "query",
            "name": "overviewId",
            "description": "overviewId",
            "allowEmptyValue": false,
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get overview list",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ApiResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "headerLanguage": {
        "name": "Accept-Language",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "default": "en",
          "enum": ["en"]
        }
      },
      "headerPlatform": {
        "name": "X-Care-Platform",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "default": "ios",
          "enum": ["ios", "android"]
        }
      },
      "headerVersion": {
        "name": "X-Care-Version",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "default": "1.0.0",
          "pattern": "^[\\d]+\\.[\\d]+\\.[\\d]+$"
        }
      },
...
    },
    "schemas": {
      "ApiResponse": {
        "required": ["success", "message", "meta"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "meta": {
            "type": "object",
            "properties": {
              "version": {
                "type": "string"
              },
              "forceUpdate": {
                "type": "boolean"
              },
              "maintenance": {
                "type": "boolean"
              },
              "hasUpdate": {
                "type": "boolean"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "responses": {
      "BadRequestError": {
        "description": "Request fails validation or doesn't meet all conditions for request",
        "content": {
          "application/json": {
            "schema": {
              "example": {
                "success": false,
                "message": "Validation failed",
                "data": ""
              },
              "allOf": [
                {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
    }
  }
}